requirements.txt
selenium==3.141.0
requests==2.25.1
PyYAML==5.4.1